chore(migration): panels-features specs to ts
