--- tests/panels-features/panels-features-fill.spec.js
@@ -1,181 +1,0 @@
-const { mainTest } = require('../../fixtures');
-const { MainPage } = require('../../pages/workspace/main-page');
-const { ColorPalettePage } = require('../../pages/workspace/color-palette-page');
-const { expect } = require('@playwright/test');
-const { random } = require('../../helpers/string-generator');
-const { TeamPage } = require('../../pages/dashboard/team-page');
-const { DashboardPage } = require('../../pages/dashboard/dashboard-page');
-const { DesignPanelPage } = require('../../pages/workspace/design-panel-page');
-const { qase } = require('playwright-qase-reporter/playwright');
-const { createTeamName } = require('../../helpers/teams/create-team-name');
-
-const teamName = createTeamName();
-
-let teamPage, mainPage, colorPalettePage, dashboardPage, designPanelPage;
-
-mainTest.beforeEach(async ({ page }) => {
-  teamPage = new TeamPage(page);
-  dashboardPage = new DashboardPage(page);
-  designPanelPage = new DesignPanelPage(page);
-  colorPalettePage = new ColorPalettePage(page);
-  mainPage = new MainPage(page);
-  await teamPage.createTeam(teamName);
-  await teamPage.isTeamSelected(teamName);
-  await dashboardPage.createFileViaPlaceholder();
-  await mainPage.isMainPageLoaded();
-});
-
-mainTest.describe(() => {
-  mainTest.beforeEach(async () => {
-    await mainPage.clickCreateBoardButton();
-    await mainPage.clickViewportTwice();
-    await mainPage.waitForChangeIsSaved();
-    await mainPage.isCreatedLayerVisible();
-  });
-
-  mainTest(qase(786, 'Add fill to board'), async () => {
-    await designPanelPage.isFillHexCodeSet('#FFFFFF');
-    await designPanelPage.isFillOpacitySet('100');
-    await expect(mainPage.viewport).toHaveScreenshot('board-fill.png', {
-      mask: mainPage.maskViewport(),
-    });
-  });
-
-  mainTest(qase(791, 'Change fill color for board'), async () => {
-    await designPanelPage.clickFillColorIcon();
-    await colorPalettePage.modalSetHex('FF0000');
-    await mainPage.clickViewportTwice();
-    await mainPage.waitForChangeIsSaved();
-    await designPanelPage.isFillHexCodeSet('#ff0000');
-    await designPanelPage.isFillOpacitySet('100');
-    await expect(mainPage.viewport).toHaveScreenshot('board-changed-fill.png', {
-      mask: mainPage.maskViewport(),
-    });
-  });
-
-  mainTest(qase(796, 'Change fill opacity for board'), async () => {
-    await designPanelPage.changeOpacityForFill('70');
-    await mainPage.clickViewportTwice();
-    await mainPage.waitForChangeIsSaved();
-    await designPanelPage.isFillHexCodeSet('#FFFFFF');
-    await designPanelPage.isFillOpacitySet('70');
-    await expect(mainPage.viewport).toHaveScreenshot('board-changed-opacity.png', {
-      mask: mainPage.maskViewport(),
-    });
-  });
-
-  mainTest(qase(811, 'Remove fill for board'), async () => {
-    await designPanelPage.clickRemoveFillButton();
-    await mainPage.waitForChangeIsSaved();
-    await expect(mainPage.viewport).toHaveScreenshot('board-removed-fill.png', {
-      mask: mainPage.maskViewport(),
-    });
-  });
-});
-
-mainTest.describe(() => {
-  mainTest.beforeEach(async () => {
-    await mainPage.createDefaultClosedPath();
-    await mainPage.isCreatedLayerVisible();
-  });
-
-  mainTest(qase(790, 'Add fill to path'), async () => {
-    await designPanelPage.clickAddFillButton();
-    await mainPage.waitForChangeIsSaved();
-    await designPanelPage.isFillHexCodeSet('#B1B2B5');
-    await designPanelPage.isFillOpacitySet('100');
-    await mainPage.waitForResizeHandlerVisible();
-    await expect(mainPage.viewport).toHaveScreenshot('path-fill.png', {
-      mask: mainPage.maskViewport(),
-    });
-  });
-
-  mainTest(qase(795, 'Change fill color for path'), async () => {
-    await designPanelPage.clickAddFillButton();
-    await mainPage.waitForChangeIsSaved();
-    await designPanelPage.clickFillColorIcon();
-    await colorPalettePage.setHex('#FF0000');
-    await mainPage.clickOnDesignTab();
-    await mainPage.waitForChangeIsSaved();
-    await designPanelPage.isFillHexCodeSet('#FF0000');
-    await designPanelPage.isFillOpacitySet('100');
-    await expect(mainPage.viewport).toHaveScreenshot('path-changed-fill.png', {
-      mask: mainPage.maskViewport(),
-    });
-  });
-
-  mainTest(qase(800, 'Change fill opacity for path'), async () => {
-    await designPanelPage.clickAddFillButton();
-    await designPanelPage.changeOpacityForFill('70');
-    await mainPage.clickOnDesignTab();
-    await mainPage.waitForChangeIsSaved();
-    await designPanelPage.isFillHexCodeSet('#B1B2B5');
-    await designPanelPage.isFillOpacitySet('70');
-    await mainPage.waitForResizeHandlerVisible();
-    await expect(mainPage.viewport).toHaveScreenshot('path-changed-opacity.png', {
-      mask: mainPage.maskViewport(),
-    });
-  });
-
-  mainTest(qase(815, 'Remove fill for path'), async () => {
-    await designPanelPage.clickAddFillButton();
-    await mainPage.waitForChangeIsSaved();
-    await designPanelPage.clickRemoveFillButton();
-    await mainPage.waitForChangeIsSaved();
-    await mainPage.clickOnDesignTab();
-    await expect(mainPage.viewport).toHaveScreenshot('path-removed-fill.png', {
-      mask: mainPage.maskViewport(),
-    });
-  });
-});
-
-mainTest.describe(() => {
-  mainTest.beforeEach(async () => {
-    await mainPage.clickCreateRectangleButton();
-    await mainPage.clickViewportTwice();
-    await mainPage.waitForChangeIsSaved();
-    await mainPage.isCreatedLayerVisible();
-  });
-
-  mainTest(qase(787, 'Add fill to shape'), async () => {
-    await designPanelPage.isFillHexCodeSet('#B1B2B5');
-    await designPanelPage.isFillOpacitySet('100');
-    await expect(mainPage.viewport).toHaveScreenshot('rectangle-fill.png', {
-      mask: mainPage.maskViewport(),
-    });
-  });
-
-  mainTest(qase(797, 'Change fill opacity for shape'), async () => {
-    await designPanelPage.changeOpacityForFill('70');
-    await mainPage.clickViewportTwice();
-    await designPanelPage.isFillHexCodeSet('#B1B2B5');
-    await designPanelPage.isFillOpacitySet('70');
-    await mainPage.waitForResizeHandlerVisible();
-    await expect(mainPage.viewport).toHaveScreenshot(
-      'rectangle-changed-opacity.png',
-      {
-        mask: mainPage.maskViewport(),
-      },
-    );
-  });
-
-  mainTest(qase(812, 'Remove fill for shape'), async () => {
-    await designPanelPage.clickRemoveFillButton();
-    await mainPage.waitForChangeIsSaved();
-    await expect(mainPage.viewport).toHaveScreenshot('rectangle--removed-fill.png', {
-      mask: mainPage.maskViewport(),
-    });
-  });
-
-  mainTest(qase(792, 'Change fill color for shape'), async () => {
-    await designPanelPage.clickFillColorIcon();
-    await colorPalettePage.setHex('#FF0000');
-    await mainPage.clickViewportTwice();
-    await mainPage.waitForChangeIsSaved();
-    await designPanelPage.isFillHexCodeSet('#FF0000');
-    await designPanelPage.isFillOpacitySet('100');
-    await expect(mainPage.viewport).toHaveScreenshot('rectangle-changed-fill.png', {
-      mask: mainPage.maskViewport(),
-    });
-  });
-});
