+++ tests/panels-features/panels-features-fill.spec.ts-snapshots/board-changed-opacity.png
