+++ tests/panels-features/panels-features-fill.spec.ts-snapshots/board-removed-fill.png
