+++ tests/panels-features/panels-features-fill.spec.ts-snapshots/path-changed-opacity.png
