+++ tests/panels-features/panels-features-fill.spec.ts-snapshots/path-fill.png
