+++ tests/panels-features/panels-features-fill.spec.ts-snapshots/path-removed-fill.png
