+++ tests/panels-features/panels-features-fill.spec.ts-snapshots/rectangle-changed-fill.png
