+++ tests/panels-features/panels-features-guides.spec.ts-snapshots/board-with-ruler-guides.png
