+++ tests/panels-features/panels-features-guides.spec.ts-snapshots/board-without-guide.png
