+++ tests/panels-features/panels-features-guides.spec.ts-snapshots/columns-guide-hide.png
