+++ tests/panels-features/panels-features-guides.spec.ts-snapshots/columns-guide-opacity-50.png
