+++ tests/panels-features/panels-features-guides.spec.ts-snapshots/rows-guide-changed-rows.png
