+++ tests/panels-features/panels-features-guides.spec.ts-snapshots/rows-guide-default.png
