+++ tests/panels-features/panels-features-guides.spec.ts-snapshots/rows-guide-opacity-50.png
