+++ tests/panels-features/panels-features-guides.spec.ts-snapshots/square-guide-unhide.png
