+++ tests/panels-features/panels-features-guides.spec.ts-snapshots/two-board-with-guide.png
