+++ tests/panels-features/panels-features-prototype.spec.ts-snapshots/add-interaction.png
