+++ tests/panels-features/panels-features-prototype.spec.ts-snapshots/connector-between-board2-and-board1.png
