+++ tests/panels-features/panels-features-prototype.spec.ts-snapshots/delete-flow.png
