+++ tests/panels-features/panels-features-prototype.spec.ts-snapshots/rename-flow.png
