+++ tests/panels-features/panels-features-zoom.spec.ts-snapshots/canvas-zoom-out.png
